+++ RELEASE_NOTES.rst
@@ -32,3 +32,4 @@ Version 1.6
 - Changed neotool text output format
 - Changed Node.is_abstract and Relationship.is_abstract from function to property
 - Removed ReadBatch.get_properties; use GraphDatabaseService.get_properties instead
+- Changed GDB.order and size to properties